fix verbosity and quiet argument handling
When missing a verbosity or quiet argument, the application was failing to run due to missing arguments that the main function was expecting. As such, the main function now only takes the target directory. Setting the log level is now handled in the entry point function, separating that from the core of the application.